PutPolicy support set -1 & version to 7.18.1 (#119)

+++ storage/token.go
@@ -91,11 +91,16 @@ type PutPolicy struct {
 	FsizeLimit int64 `json:"fsizeLimit,omitempty"`
 
 	// 开启 MimeType 侦测功能，并按照下述规则进行侦测；如不能侦测出正确的值，会默认使用 application/octet-stream 。
-	// 设为非 0 值，则忽略上传端传递的文件 MimeType 信息，并按如下顺序侦测 MimeType 值：
-	// 1. 侦测内容； 2. 检查文件扩展名； 3. 检查 Key 扩展名。
-	// 默认设为 0 值，如上传端指定了 MimeType 则直接使用该值，否则按如下顺序侦测 MimeType 值：
-	// 1. 检查文件扩展名； 2. 检查 Key 扩展名； 3. 侦测内容。
-	DetectMime uint8 `json:"detectMime,omitempty"`
+	// 默认设为 0 时：如上传端指定了 MimeType 则直接使用该值，否则按如下顺序侦测 MimeType 值：
+	//		1. 检查文件扩展名；
+	//		2. 检查 Key 扩展名；
+	//		3. 侦测内容。
+	// 设为 1 时：则忽略上传端传递的文件 MimeType 信息，并按如下顺序侦测 MimeType 值：
+	//		1. 侦测内容；
+	//		2. 检查文件扩展名；
+	//		3. 检查 Key 扩展名。
+	// 设为 -1 时：无论上传端指定了何值直接使用该值。
+	DetectMime int `json:"detectMime,omitempty"`
 
 	// 限定用户上传的文件类型。指定本字段值，七牛服务器会侦测文件内容以判断 MimeType，再用判断值跟指定值进行匹配，匹配成功则允许上传，匹配失败则返回 403 状态码。示例：
 	// image/* 表示只允许上传图片类型
